Crouch and Fixed Body Rotation
Added a crouching feature. It stops headbobbing and sprinting while active. Camera is built-in to the head via Cinemachine follow and the head "ducks" down when crouching. This is achieved by lerping between the original height and the desired crouch height.
I also fixed and improved the body rotation. It was not rotating with the camera before, but I took the opportunity to have the body slerp towards the camera's forward vector, since the body and "head" are so nicely decoupled already. The rotation quickens based on the player's speed, so sprint will rotate faster than a crouched stance.