NPC/Object interact + head look rotations
Implemented most of the structure I need to make interactions with npcs and objects via UI and player input. Added Unity Animation Rigging to both player and NPC to rotate head towards the player when in vicinity and to have the player's head rotate in the direction of the camera.

Issuing a stern reminder to myself for later: make sure the final humanoid models are using the proper left-hand rotation model that Unity uses. Working around it is misery.