Interaction popup added
UI element is enabled or disabled when interacting with an npc within interaction range. The player can now see a popup appear to notify the player that they are within range to interact. Also added the ability to determine which interactable objects are closest to the player to avoid any popup overlapping when two or more objects are within reach