Refactor Interactions
Refactored the interaction system by using Interfaces. The benefit is that I can use a generic interface object to call to unique monobehavior methods that all perform different behaviors in the game. Plus, it adds a layer of consistency by forcing method names to fall under the same name and be included in a class where it is needed.